cpuidle: set the current wake up latency Date	Fri, 18 Oct 2013 13:52:27 +0200
Save the current wake up latency of a core. This latency is not always
the latency of a defined c-state but can also be an intermediate value
when a core is ready to shutdown (timer migration, cache flush ...) but
wait for the last core of the cluster to finalize the cluster power down.
This latter use case is not manage by the current version of the patch because
it implies that the cpuidle drivers set the wake up latency instead of the
cpuidle core.

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>

Signed-off-by: Paul Reioux <reioux@gmail.com>

Conflicts:
	drivers/cpuidle/cpuidle.c

Signed-off-by: XileForce <rdaguiar360@gmail.com>

Original commit: https://github.com/XileForce/Vindicator-S6/commit/56e4dbeafc64f2eb77c261576f8a8323fb505b6a