arm64: refactoring cpu reset routines code
Add  a simplified version of cpu_soft_restart() with some changes needed
for kexec in the new files cpu_reset.S, and cpu_reset.h.

When a CPU is reset it needs to be put into the exception level it had when
it entered the kernel. Update cpu_soft_restart() to accept an argument
which signals if the reset address should be entered at EL1 or EL2, and
add a new hypercall HVC_SOFT_RESTART which is used for the EL2 switch.

Based on commits:

 https://github.com/hizukiayaka/linux-kernel/commit/68234df4ea7939f98431aa81113fbdce10c4a84b
 https://github.com/hizukiayaka/linux-kernel/commit/f9076ecfb1216a478312b1c078d04792df6d4477
 https://github.com/hizukiayaka/linux-kernel/commit/e7227d0e528f9a96d4a866f43e20dd9b33f0e782