[for-next,3/4] ring-buffer: Allways do the trace_recursive checks
From: "Steven Rostedt (Red Hat)" <rostedt@goodmis.org>

Currently the trace_recursive checks are only done if CONFIG_TRACING
is enabled. That was because there use to be a dependency with tracing
for the recursive checks (it used the task_struct trace recursive
variable). But now it uses its own variable and there is no dependency.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>

Patch: https://patchwork.kernel.org/patch/6495551/