mptcp: Avoid reinjecting on same subflow
It may be that we call mptcp_reinject_data on a subflow, which is still
eligible for sending (e.g., when receiving a REMOVE_ADDR, we transition
the subflow to CLOSE only after calling mptcp_reinject_data).
In this case it is possible that we reinject the data on the same
subflow again.

We have to prevent this, by setting tp->pf before calling
mptcp_push_pending_frames. tp->pf set to 1 will make sure that data does
not get send on this subflow.

Might be related to #100

Fixes: Zero-day bug
Signed-off-by: Christoph Paasch <christoph.paasch@gmail.com>

Original commit: https://github.com/multipath-tcp/mptcp/commit/6888208d1177b05cda6eb1477f4970a88acb87e4