mptcp: Correctly handle overlapping incoming data
It may well be, that a peer is retransmitting old data together with new
data (e.g., if a peer would do Redundant Data Bundling).

So, we have to handle this case gracefully, where seq < copied_seq.

This patch handles it in the regular as well as in the fallback-case.

Fixes: 7208b33 (Fix: Correctly handle malformed DSS-mappings)
Signed-off-by: Christoph Paasch <christoph.paasch@gmail.com>

Based on commit: https://github.com/multipath-tcp/mptcp/commit/b4be1cf756b03621f7a2c11491de02617cf8bbeb