mptcp: Fix static-key handling in interrupt context
static_key_slow_inc cannot be called in interrupt context. Because, it
takes a mutex_lock and as this one might sleep, we cannot do this in
interrupt context.

We thus try in this patch to implement something similar to
b90e579 (net: dont call jump_label_dec from irq context).

The big difference is that in MPTCP, the label can be inc'd not only by
an app creating an MPTCP-connection, but also by the kernel when
accepting new subflows (we need to inc it per subflow). Thus, the
deferred variable's meaning is inversed wrt to b90e579. In case it
is > 0, that means we have missed an inc.

We must inc the jump-label in mptcp_init_tcp_sock, because this is the only
place where we are guaranteed to be in user-context. That way, the label
gets enabled when on the client-side we create the socket and on the
server-side when the listening-socket is created. If we would not have
done this, then label would never be inc'd on the server-side because
meta-creation is done in interrupt-context and thus the inc of the flag
is deferred.

Cc: Baptiste Jonglez <baptiste@bitsofnetworks.org>
Reported-by: Baptiste Jonglez <baptiste@bitsofnetworks.org>
Fixes: 1ec7a15 (mptcp: Add static-key check in mptcp())
Signed-off-by: Christoph Paasch <christoph.paasch@gmail.com>
Signed-off-by: Sébastien Barré <sebastien.barre@uclouvain.be>

Original commit: https://github.com/multipath-tcp/mptcp/commit/3cdd069f255db4484863bc1a6f1e1205b8eeb50c