defconfig: Use zcache instead zswap
Zcache  is a backend for the Transcendent memory system. Transcendent
memory provides a RAM-like memory that can only be accessed a page at
a time by using put and get calls. This is unlike normal memory that
can be accessed a byte at a time. The frontswap and cleancache systems
hook attempts to swap and reclaim file-system page caches respectively
and send them to the transcendent memory backends. When zcache is used
as a backend, the data is compressed and stored in the RAM. When it
fills up, compressed pages are evicted to the swap. (an alternate
backend is RAMster which shares a pool of RAM across networked computers).
Using only the frontswap frontend with the zcache backend works just
like zswap. (In fact zswap is a simplified subset of zcache).

Signed-off-by: Alexander Alexeev <alexax66@gmail.com>