arm64: debug: report TRAP_TRACE instead of TRAP_HWBRPT for singlesteplogin
Single-step traps to userspace (e.g. via ptrace) are expected to use
the TRAP_TRACE for the si_code field of the siginfo, as opposed to
TRAP_HWBRPT that we report currently.

Fix the reported value, which has no effect on existing and legacy
builds of GDB.

Reported-by: Yao Qi <yao.qi@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>

Based on patch: https://patchwork.kernel.org/patch/9309541/