[v2,1/2] ftrace: Provide API to use global filtering for ftrace ops
Currently the global_ops filtering hash is not available to outside
users registering for function tracing. Provide an API for those
users to be able to choose global filtering.

This is in preparation for pstore's ftrace feature to be able to
use the global filters.

v2 changes: don't set global filter if ops were initialized before.

Suggested-by: Steven Rostedt <rostedt@goodmis.org>
Cc: Anton Vorontsov <anton@enomsg.org>
Cc: Colin Cross <ccross@android.com>
Cc: Kees Cook <keescook@chromium.org>
Cc: Tony Luck <tony.luck@intel.com>
Signed-off-by: Joel Fernandes <joelaf@google.com>

Used patch: https://patchwork.kernel.org/patch/9430471/