[v2,2/2] pstore: Use global ftrace filters for function trace filtering
Currently, pstore doesn't have any filters setup for function tracing.  This
has the associated overhead and may not be useful for users looking for tracing
specific set of functions.

ftrace's regular function trace filtering is done writing to
tracing/set_ftrace_filter however this is not available if not requested.
Inorder to be able to use this feature, the support to request global filtering
introduced earlier in the series should be requested before registering the
ftrace ops. Here we do the same.

Signed-off-by: Joel Fernandes <joelaf@google.com>

Used patch: https://patchwork.kernel.org/patch/9430473/